transfer screen open analytics to presenter
