+++ app/src/main/java/org/stepik/android/presentation/profile/ProfilePresenter.kt
@@ -1,8 +1,11 @@
 package org.stepik.android.presentation.profile
 
+import android.os.Bundle
 import io.reactivex.Scheduler
 import io.reactivex.rxkotlin.plusAssign
 import io.reactivex.rxkotlin.subscribeBy
+import org.stepic.droid.analytic.AmplitudeAnalytic
+import org.stepic.droid.analytic.Analytic
 import org.stepic.droid.di.qualifiers.BackgroundScheduler
 import org.stepic.droid.di.qualifiers.MainScheduler
 import org.stepik.android.domain.profile.interactor.ProfileInteractor
@@ -13,6 +16,7 @@ class ProfilePresenter
 @Inject
 constructor(
     private val profileInteractor: ProfileInteractor,
+    private val analytic: Analytic,
 
     @BackgroundScheduler
     private val backgroundScheduler: Scheduler,
@@ -46,6 +50,7 @@ constructor(
                         if (profileData.isCurrentUser && profileData.user.isGuest) {
                             ProfileView.State.EmptyLogin
                         } else {
+
                             ProfileView.State.Content(profileData)
                         }
                 },
@@ -53,6 +58,8 @@ constructor(
                     val oldState = state
                     if (oldState !is ProfileView.State.Content) {
                         state = ProfileView.State.Empty
+                    } else {
+                        sendScreenOpenEvent(oldState.profileData.isCurrentUser)
                     }
                 },
                 onError = {
@@ -70,5 +77,21 @@ constructor(
             ?.user
             ?: return
         view?.shareUser(user)
+    }
+
+    private fun sendScreenOpenEvent(isCurrentUser: Boolean) {
+        val state = if (isCurrentUser) {
+            AmplitudeAnalytic.Profile.Values.SELF
+        } else {
+            AmplitudeAnalytic.Profile.Values.OTHER
+        }
+
+        analytic.reportAmplitudeEvent(
+            AmplitudeAnalytic.Profile.PROFILE_SCREEN_OPENED, mapOf(
+                AmplitudeAnalytic.Profile.Params.STATE to state
+            ))
+        analytic.reportEvent(Analytic.Profile.PROFILE_SCREEN_OPENED, Bundle().apply {
+            putString(Analytic.Profile.Params.STATE, state)
+        })
     }
 }
